travis: Pin 0.9.x release branch to current stable but with version number
This patch does not land on master intentionally.
We pin the releases now on versions (as in "version string") for the
rust compiler because the latest release does not build with beta either
(it was released because the previous release did not build on beta
either).

This problem is then fixed and if there will be a v0.9.4 (which I hope
not), it will build on travis.

Should've done this way sooner.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>